% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyboxplot.R
\name{shinyboxplot}
\alias{shinyboxplot}
\title{Shiny boxplot App}
\usage{
shinyboxplot()
}
\value{
Different boxplots and Statistic summary based on t.test type
}
\description{
This function calls the Shiny Web Application on browser
}
\details{
Use this function to display the boxplots and statistic summary on web browser
}
\examples{
\dontrun{shinyboxplot()}
}
\keyword{boxplot}
\keyword{geom_boxplot}
\keyword{shiny}
