% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myttest.R
\name{myttest}
\alias{myttest}
\title{myttest function}
\usage{
myttest(x, y, alpha, paired = FALSE)
}
\arguments{
\item{x}{numeric x value}

\item{y}{numeric y value}

\item{alpha}{numeric alpha value}

\item{paired}{logical value for paired parameter, TRUE or FALSE}
}
\value{
a list of different t.test type, p value
and boxplots for different test type
}
\description{
Test the NULL hypothesis using t.test theory and make some boxplots
}
\examples{

set.seed(32); x=rnorm(30,mean=10,sd=15)
set.seed(35); y=rnorm(30,mean=8,sd=15)
ans1=myttest(x,y,alpha=0.05,paired=FALSE)
print(ans1)
class(ans1)
plot(ans1)
}
